# docs/swagger.yaml
definitions:
...
  getMenu.Item:
    properties:
      available:
        type: boolean
      description:
        example: burger with cheese
        type: string
      name:
        example: Cheeseburger
        type: string
      price:
        example: 122
        type: number
    type: object
  getMenu.Response:
    properties:
      menu:
        items:
          $ref: '#/definitions/getMenu.Item'
        type: array
      restaurant_id:
        example: 1
        type: integer
    type: object
...
  getOrdersForUser.Response:
    properties:
      ordersStruct:
        items:
          $ref: '#/definitions/ordersStruct.Order'
        type: array
    type: object
...
  getRestaurantByID.Response:
    properties:
      menu_items:
        items:
          $ref: '#/definitions/getRestaurantByID.item'
        type: array
      restaurant_address:
        example: 112 address
        type: string
      restaurant_id:
        example: 14
        type: integer
      restaurant_name:
        example: mac
        type: string
      restaurant_phone:
        example: "89053435656"
        type: string
    type: object
  getRestaurantByID.item:
    properties:
      item_description:
        example: burger with cheese
        type: string
      item_id:
        example: 1
        type: integer
      item_name:
        example: cheeseburger
        type: string
      item_price:
        example: 122
        type: number
    type: object
...
  newMenuItem.Request:
    properties:
      available:
        type: boolean
      description:
        example: burger with beef
        type: string
      name:
        example: Burger
        type: string
      price:
        example: 52.2
        type: number
    type: object
  newMenuItem.Response:
    properties:
      available:
        type: boolean
      description:
        example: burger with beef
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Burger
        type: string
      price:
        example: 52.2
        type: number
      restaurant_id:
        example: 1
        type: integer
    type: object
...
  ordersStruct.Item:
    properties:
      item_name:
        example: Burger with cheese
        type: string
      item_price:
        example: 100
        type: number
      menu_item_id:
        example: 1
        type: integer
      quantity:
        example: 3
        type: integer
    type: object
  ordersStruct.Order:
    properties:
      created_at:
        example: "2013-08-20T18:08:41+00:00"
        type: string
      delivery_Address:
        example: 1223 address
        type: string
      items:
        items:
          $ref: '#/definitions/ordersStruct.Item'
        type: array
      restaurant_Address:
        example: 123 address
        type: string
      restaurant_Name:
        example: Mac
        type: string
      restaurant_Phone:
        example: "89056666666"
        type: string
      status:
        example: pending
        type: string
      total_price:
        example: 300
        type: number
      user_name:
        example: Ivan
        type: string
    type: object
...
  placeorder.Request:
    properties:
      address:
        example: 123 address
        type: string
      items:
        items:
          properties:
            menuitem_id:
              example: 6
              type: integer
            quantity:
              example: 5
              type: integer
          type: object
        type: array
      restaurant_id:
        example: 14
        type: integer
    type: object
  placeorder.Response:
    properties:
      created_at:
        example: Tue, 17 Jun 2025 00:25:16 +0000
        type: string
      items:
        items:
          properties:
            menuitem_id:
              example: 6
              type: integer
            quantity:
              example: 5
              type: integer
          type: object
        type: array
      order_id:
        example: 12
        type: integer
      restaurant_id:
        example: 14
        type: integer
      status:
        example: pending
        type: string
      user_address:
        example: 123 address
        type: string
    type: object
...
  response.Response:
    properties:
      error:
        example: error message
        type: string
    type: object
info:
  contact: {}
  description: REST API for food delivery
  title: GodFood API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию по заказам для авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно получен
          schema:
            $ref: '#/definitions/getOrdersForUser.Response'
        "204":
          description: Нет заказов
        "400":
          description: Неккоректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Доступ Запрещен
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение заказов по JWT
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Создает новый заказ
      parameters:
      - description: Данные для добавления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/placeorder.Request'
      produces:
      - application/json
      responses:
        "200":
          description: 'Новый заказ успешно создан '
          schema:
            $ref: '#/definitions/placeorder.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Серверная Ошибка
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание нового заказа авторизованным пользователем
      tags:
      - Orders
...
  /restaurants/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию по ресторану(Айди, имя, адрес, телефон)
        и меню по айди
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ресторан успешно получен
          schema:
            $ref: '#/definitions/getRestaurantByID.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Ресторан не найден
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получение Ресторана по айди
      tags:
      - Restaurants
  /restaurants/{id}/menuItems:
    get:
      consumes:
      - application/json
      description: Возвращает меню ресторана по айди
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Меню успешно получено
          schema:
            $ref: '#/definitions/getMenu.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Ресторан не найден
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получение меню по айди
      tags:
      - Restaurants
  /restaurants/menuItems:
    post:
      consumes:
      - application/json
      description: Создает новую позицию в меню рессторана по JWT
      parameters:
      - description: Данные для добавления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/newMenuItem.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Новая позиция успешно добавлена
          schema:
            $ref: '#/definitions/newMenuItem.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Серверная Ошибка
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Добавление новой позиции в меню
      tags:
      - Restaurants
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"